+++ types.h
@@ -1,5 +1,5 @@
 typedef struct {
-	int type;	//1 pour client, 2 pour cuisinier
+	long type;	//1 pour client, 2 pour cuisinier
 	int serveur;
 	int choix;
 	pid_t expediteur;
@@ -7,14 +7,14 @@ typedef struct {
 command_t;
 
 typedef struct {
-	int type;	//3 pour serveur, 2 pour cuisinier
+	long type;	//3 pour serveur, 2 pour cuisinier
 	int choix;
 	pid_t expediteur;
 }
 commandcuiserv_t;
 
 typedef struct {
-	int type;
+	long type;
 	int choix;
 } 
 commandcliserv_t;
